+++ totalRP3/Locales/enUS.lua
@@ -1599,6 +1599,28 @@ If you wish to report %s's profile and you cannot target them you will need to o
 	AUTOMATION_TEST_BUTTON = "Test condition",
 	AUTOMATION_TEST_HELP = "Executes the supplied macro condition, printing the option it would select to the chat frame.",
 	AUTOMATION_TEST_OUTPUT = "Test condition result: %s",
+
+	WHATS_NEW_26_0 = [[# Changelog version 2.5.0
+
+## Added
+
+- Added 503 icons, 15 images and 70 musics from patch 10.1.
+- Added Automation system, accessible in the settings
+  - This allows you to automate actions based on macro conditionals.
+  - Actions include changing your profile, your roleplay status, whether the map scan is enabled or not...
+  - For a list of macro conditionals, see [our wiki article](https://github.com/Total-RP/Total-RP-3/wiki/Analytics-in-Total-RP-3).
+- Added Analytics module
+  - **This is strictly opt-in and only applies to users of the Wago Addons client who enabled "Help addon developers".**
+  - For more details, see [our wiki article](https://github.com/Total-RP/Total-RP-3/wiki/Analytics-in-Total-RP-3).
+- Added entry in the new addon compartment located below the calendar button, with the same actions as the minimap button.
+- Added tooltip indicator if the guild fields have been customized, as well as options to display both/either custom and in-game guild info.
+- Added ability to get the party started.
+
+## Changed
+
+- "Always show target" nameplate option will now display the target nameplate even if it is OOC and "Hide out of character units" was enabled.
+
+]],
 };
 
 -- Bindings and FrameXML Global Strings