+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# Changelog version 2.5.0
+
+## Added
+
+- Added 503 icons, 15 images and 70 musics from patch 10.1.
+- Added Automation system, accessible in the settings
+  - This allows you to automate actions based on macro conditionals.
+  - Actions include changing your profile, your roleplay status, whether the map scan is enabled or not...
+  - For a list of macro conditionals, see [our wiki article](https://github.com/Total-RP/Total-RP-3/wiki/Analytics-in-Total-RP-3).
+- Added Analytics module
+  - **This is strictly opt-in and only applies to users of the Wago Addons client who enabled "Help addon developers".**
+  - For more details, see [our wiki article](https://github.com/Total-RP/Total-RP-3/wiki/Analytics-in-Total-RP-3).
+- Added entry in the new addon compartment located below the calendar button, with the same actions as the minimap button.
+- Added tooltip indicator if the guild fields have been customized, as well as options to display both/either custom and in-game guild info.
+- Added ability to get the party started.
+
+## Changed
+
+- "Always show target" nameplate option will now display the target nameplate even if it is OOC and "Hide out of character units" was enabled.
+
 # Changelog version 2.4.6
 
 ## Added